created abstractEntity, updated FuelTypes enum, set all id to Long type

+++ src/main/java/pl/coderslab/servicestation/repositories/PartRepository.java
@@ -3,6 +3,6 @@ package pl.coderslab.servicestation.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.coderslab.servicestation.models.Part;
 
-public interface PartRepository extends JpaRepository<Part, Integer> {
+public interface PartRepository extends JpaRepository<Part, Long> {
 
 }
